+++ airflow/contrib/sensors/sagemaker_transform_sensor.py
@@ -16,47 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sagemaker_transform`."""
 
-from airflow.contrib.hooks.sagemaker_hook import SageMakerHook
-from airflow.providers.amazon.aws.sensors.sagemaker_base import SageMakerBaseSensor
-from airflow.utils.decorators import apply_defaults
+import warnings
 
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.sensors.sagemaker_transform import SageMakerTransformSensor  # noqa
 
-class SageMakerTransformSensor(SageMakerBaseSensor):
-    """
-    Asks for the state of the transform state until it reaches a terminal state.
-    The sensor will error if the job errors, throwing a AirflowException
-    containing the failure reason.
-
-    :param job_name: job_name of the transform job instance to check the state of
-    :type job_name: str
-    """
-
-    template_fields = ['job_name']
-    template_ext = ()
-
-    @apply_defaults
-    def __init__(self,
-                 job_name,
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.job_name = job_name
-
-    def non_terminal_states(self):
-        return SageMakerHook.non_terminal_states
-
-    def failed_states(self):
-        return SageMakerHook.failed_states
-
-    def get_sagemaker_response(self):
-        sagemaker = SageMakerHook(aws_conn_id=self.aws_conn_id)
-
-        self.log.info('Poking Sagemaker Transform Job %s', self.job_name)
-        return sagemaker.describe_transform_job(self.job_name)
-
-    def get_failed_reason_from_response(self, response):
-        return response['FailureReason']
-
-    def state_from_response(self, response):
-        return response['TransformJobStatus']
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sagemaker_transform`.",
+    DeprecationWarning, stacklevel=2
+)
