+++ docs/operators-and-hooks-ref.rst
@@ -384,7 +384,7 @@ These integrations allow you to perform various operations within the Amazon Web
      - :mod:`airflow.providers.amazon.aws.sensors.sagemaker_base`,
        :mod:`airflow.providers.amazon.aws.sensors.sagemaker_endpoint`,
        :mod:`airflow.providers.amazon.aws.sensors.sagemaker_training`,
-       :mod:`airflow.contrib.sensors.sagemaker_transform_sensor`,
+       :mod:`airflow.providers.amazon.aws.sensors.sagemaker_transform`,
        :mod:`airflow.contrib.sensors.sagemaker_tuning_sensor`
 
    * - `Amazon Simple Notification Service (SNS) <https://aws.amazon.com/sns/>`__
