+++ scripts/ci/pylint_todo.txt
@@ -95,7 +95,7 @@
 ./airflow/providers/amazon/aws/sensors/sagemaker_base.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_endpoint.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_training.py
-./airflow/contrib/sensors/sagemaker_transform_sensor.py
+./airflow/providers/amazon/aws/sensors/sagemaker_transform.py
 ./airflow/contrib/sensors/sagemaker_tuning_sensor.py
 ./airflow/providers/microsoft/azure/sensors/wasb.py
 ./airflow/sensors/weekday_sensor.py
